Rename module loaders to loader
