+++ pytissueoptics/scene/__init__.py
@@ -2,4 +2,4 @@ from .solids import Cuboid, Cube, Sphere
 from .geometry import Vector
 from .materials import Material
 from .scene import Scene
-from .loaders import Loader
+from .loader import Loader
