+++ pytissueoptics/scene/loader/loader.py
