+++ pytissueoptics/scene/loader/parsers/parser.py
